rename templatetag, add counting recipe in basket
