+++ templates/index.html
@@ -1,4 +1,4 @@
-{% load recipes_filters %}
+{% load recipes_tools %}
 <!doctype html>
 <html lang="en">
     <head>
