+++ templates/nav.html
@@ -1,3 +1,4 @@
+{% load recipes_tools %}
 <header class="header">
     <nav class="nav">
         <div class="nav__container container">
@@ -8,7 +9,7 @@
                 <li class="nav__item"><a href="{% url 'create_recipe' %}" class="nav__link link">Создать рецепт</a></li>
                 <li class="nav__item"><a href="#" class="nav__link link">Избранное</a></li>
                 {% endif %}
-                <li class="nav__item"><a href="#" class="nav__link link">Список покупок</a> <span class="badge badge_style_blue nav__badge" id="counter">4</span></li>
+                <li class="nav__item"><a href="#" class="nav__link link">Список покупок</a> <span class="badge badge_style_blue nav__badge" id="counter">{{ request.user|count_purchases }}</span></li>
 
             </ul>
             <ul class="nav__items list">
