+++ templates/viewRecipe.html
@@ -1,5 +1,5 @@
 {% load static %}
-{% load recipes_filters %}
+{% load recipes_tools %}
 <!doctype html>
 <html lang="en">
 <head>
